moved shim strings to shim files
