+++ packages/agent-utils/src/scripts/js-engine/shims/globals/math.ts
@@ -3,3 +3,10 @@ export const mathShim = {
     return __wrap_subinvoke("plugin/math", "random", {}).value;
   }
 };
+
+// language=JavaScript
+module.exports = `var mathShim = {
+    random: function () {
+        return __wrap_subinvoke("plugin/math", "random", {}).value;
+    }
+};`
